// src/interfaces/Reservations.tsx
export interface ReserveInfo {
    bookId: number;
    campId: number;
    siteId: number;
    hostId: number;
    userId: number;
    checkInDate: string;
    checkOutDate: string;
    Camp_checkIn: string;
    Camp_checkOut: string;
    adults: number;
    children: number;
    siteDesc: string;
    siteInfo: string;
    siteMainImage: string;
    siteName: string;
    sitePrice: number;

}